full-application/environment-visual/requirements.txt
# Fish Audio and Gemini API dependencies for environment-visual
fishaudio
python-dotenv
google-genai
Pillow
